+++ realtime-system/supplements/inhouse-ribbon-calibration/gencal/src/interpol.h
@@ -15,13 +15,13 @@ namespace Interpol
   // Data structure containing the lookup and interpolation info
   typedef struct
   {
-    int points;         ///< number of points in the table, for both arrays !
-    int32_t *x_values;  ///< x-values (input) must be sorted / monotonic !!
-    int32_t *y_values;  ///< y-values (output) can be arbitrary
+    int points;       ///< number of points in the table, for both arrays !
+    float *x_values;  ///< x-values (input) must be sorted / monotonic !!
+    float *y_values;  ///< y-values (output) can be arbitrary
   } InterpolData;
 
   // Table lookup with interpolation
-  int32_t InterpolateValue(InterpolData *table, int32_t x);
+  float InterpolateValue(InterpolData *table, float x);
 }  // namespace
 
 //EOF
